backend/requirement.txt
fastapi
uvicorn
watchdog
langchain
sentence-transformers
faiss-cpu
httpx
python-dotenv
